Recombine ConsentSection
